requirements.txt
Django==3.2.7
Pillow==8.4.0
pytesseract==0.3.8
